工作流节点列表接口修改

+++ rosettaflow-mapper/src/main/java/com/platon/rosettaflow/dto/AlgorithmDto.java
@@ -61,7 +61,7 @@ public class AlgorithmDto {
     /**
      * 所需的核数 (单位: 个)
      */
-    private Long costCpu;
+    private Integer costCpu;
 
     /**
      * GPU核数(单位：核)
